% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_sheets.R
\name{add_sentrix_key}
\alias{add_sentrix_key}
\title{Build a Sentrix key column for matching}
\usage{
add_sentrix_key(df)
}
\description{
Tries (in order):
Basename
Sentrix_ID + "\emph{" + Sentrix_Position
Sentrix_Barcode + "}" + Sentrix_Position
Returns the data.frame with a new \verb{_sentrix_key} column. Rows that can't
be keyed get NA.
}
\keyword{internal}
