% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_sheets.R
\name{idat_basenames}
\alias{idat_basenames}
\title{Reduce a list of IDAT paths to unique Sentrix basenames}
\usage{
idat_basenames(idat_paths)
}
\description{
"206203800149_R01C01_Grn.idat.gz" -> "206203800149_R01C01"
Returns a data.frame mapping basename -> full Grn / Red file paths.
}
\keyword{internal}
