% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_processing.R
\name{load_input_data}
\alias{load_input_data}
\title{Load beta values from various input formats}
\usage{
load_input_data(input, n_cores, verbose)
}
\description{
Side effects when input is an IDAT directory: stashes a discovered pheno
frame in \code{.qc_env$discovered_pheno}, available for use by clocker()
when the user didn't supply one explicitly.
}
\keyword{internal}
