% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_sheets.R
\name{looks_like_sample_sheet}
\alias{looks_like_sample_sheet}
\title{Test whether a CSV file looks like an Illumina-style sample sheet}
\usage{
looks_like_sample_sheet(path)
}
\description{
Test whether a CSV file looks like an Illumina-style sample sheet
}
\keyword{internal}
